just some further testing and deplicating the dt rounding structs
