+++ crates/ryo3-jiff/src/internal/mod.rs
@@ -1,3 +1,5 @@
+mod datetime_round;
 mod span_relative_to;
 
-pub(crate) use span_relative_to::*;
+pub use datetime_round::*;
+pub use span_relative_to::*;
