+++ crates/ryo3-jiff/src/ry_zoned.rs
@@ -1,5 +1,5 @@
 use crate::delta_arithmetic_self::RyDeltaArithmeticSelf;
-use crate::dev::{JiffUnit, RyDateTimeRound};
+use crate::internal::IntoDateTimeRound;
 use crate::pydatetime_conversions::zoned_to_pyobject;
 use crate::ry_datetime::RyDateTime;
 use crate::ry_span::RySpan;
@@ -7,7 +7,7 @@ use crate::ry_time::RyTime;
 use crate::ry_timestamp::RyTimestamp;
 use crate::ry_timezone::RyTimeZone;
 use crate::RyDate;
-use jiff::{Zoned, ZonedRound};
+use jiff::Zoned;
 use pyo3::basic::CompareOp;
 use pyo3::types::{PyDateTime, PyType};
 use pyo3::{pyclass, pymethods, Bound, FromPyObject, IntoPyObject, PyAny, PyErr, PyResult, Python};
@@ -241,7 +241,7 @@ impl RyZoned {
         RyTimeZone::from(self.0.time_zone())
     }
 
-    fn round(&self, option: IntoZonedRound) -> PyResult<Self> {
+    fn round(&self, option: IntoDateTimeRound) -> PyResult<Self> {
         self.0
             .round(option)
             .map(RyZoned::from)
@@ -312,25 +312,6 @@ impl Display for RyZoned {
 impl From<Zoned> for RyZoned {
     fn from(value: Zoned) -> Self {
         RyZoned(value)
-    }
-}
-
-#[derive(Debug, Clone, FromPyObject)]
-enum IntoZonedRound {
-    RyDateTimeRound(RyDateTimeRound),
-    JiffUnit(JiffUnit),
-}
-
-impl From<IntoZonedRound> for ZonedRound {
-    fn from(val: IntoZonedRound) -> Self {
-        match val {
-            // TODO: this is ugly
-            IntoZonedRound::RyDateTimeRound(round) => ZonedRound::new()
-                .smallest(round.smallest.0)
-                .mode(round.mode.0)
-                .increment(round.increment),
-            IntoZonedRound::JiffUnit(unit) => unit.0.into(),
-        }
     }
 }
